requirements.txt
win10toast
gTTS
playsound
pyttsx3
face_recognition
numpy
emoji
mysql.connector
requests
tabulate